requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
pydantic==2.5.1
python-multipart==0.0.6
requests==2.31.0
pandas==2.1.0
numpy==1.26.0
python-magic==0.4.27
chromadb==0.4.15
openai==1.3.0
